+++ src/main/java/com/cinchapi/ccl/syntax/AbstractSyntaxTree.java
@@ -67,4 +67,25 @@ public interface AbstractSyntaxTree {
      */
     public Symbol root();
 
+    /**
+     * This method is part of the visitor design pattern It accepts an
+     * {@link Visitor} as {@code visitor}.
+     *
+     * @param visitor the visitor
+     * @return the data
+     */
+    public default <T> T accept(Visitor<T> visitor) {
+        return accept(visitor, new Object[] {});
+    }
+
+    /**
+     * This method is part of the visitor design pattern It accepts an
+     * {@link Visitor} as {@code visitor} and some data {@code data}.
+     *
+     * @param visitor the visitor
+     * @param data the data
+     * @return the data
+     */
+    public <T> T accept(Visitor<T> visitor, Object... data);
+    
 }